remove column select from topbar
